refac: move const and utils
