+++ src/main.rs
@@ -5,7 +5,8 @@ use plonky2::plonk::circuit_builder::CircuitBuilder;
 use plonky2::plonk::circuit_data::CircuitConfig;
 use plonky2::plonk::config::{GenericConfig, PoseidonGoldilocksConfig};
 use plonky2::util::timing::TimingTree;
-use plonky2_sha256::circuit::{array_to_bits, make_circuits};
+use plonky2_sha256::circuit::make_circuits;
+use plonky2_sha256::utils::array_to_bits;
 use sha2::{Digest, Sha256};
 
 pub fn prove_sha256(msg: &[u8]) -> Result<()> {
